Move recover panic handling logic to option
